# hospital-server/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  dto.CreateCommentReq:
    properties:
      content:
        type: string
      is_internal:
        type: boolean
    required:
    - content
    type: object
...
  dto.CreateTicketReq:
    properties:
      assignee_id:
        type: string
      description:
        type: string
      hospital_id:
        type: string
      priority:
        type: integer
      title:
        maxLength: 200
        type: string
      type_id:
        type: string
    required:
    - title
    - type_id
    type: object
...
  dto.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  portal.UpdateProfileReq:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      phone:
        type: string
      real_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 医院信息系统 API
  title: HIS API
  version: "1.0"
paths:
...
  /api/portal/v1/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 获取个人资料
      tags:
      - portal-profile
    put:
      consumes:
      - application/json
      parameters:
      - description: 资料信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/portal.UpdateProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 更新个人资料
      tags:
      - portal-profile
  /api/portal/v1/tickets:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PageResult'
              type: object
      security:
      - BearerAuth: []
      summary: 门户工单列表
      tags:
      - portal-tickets
    post:
      consumes:
      - application/json
      parameters:
      - description: 工单信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTicketReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 门户创建工单
      tags:
      - portal-tickets
  /api/portal/v1/tickets/{id}:
    get:
      parameters:
      - description: 工单 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 门户获取工单
      tags:
      - portal-tickets
  /api/portal/v1/tickets/{id}/comments:
    post:
      consumes:
      - application/json
      parameters:
      - description: 工单 ID
        in: path
        name: id
        required: true
        type: string
      - description: 评论内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommentReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 门户添加评论
      tags:
      - portal-tickets
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token (format: Bearer {token})'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"